Move request response information from logger to request trait and return it as string
